Increase onedriveb default chunk size to 100MB
If the upload script is run on a relatively fast internet connection,
then a low chunk size can give a bad upload speed, since we have to
request for a new access token between every chunk.

The chunk size should of course be tailored to a particular internet
connection. So increasing the default chunk size doesn't bring any
harm, and can also make it less likely for the user to pick a value
that is too low, which leading to a bad upload performance.